Add files via upload
A major online retailer has approached us with a concern regarding customer retention. Despite having a vast customer base, they're facing difficulties in keeping them, with about 90% of customers not returning after their initial purchase.

Our primary objective, as mandated by the company's leadership, is to develop a predictive model that can accurately identify customers who are unlikely to make repeat purchases. Additionally, the leadership is interested in understanding the underlying factors that contribute to a customer's decision not to reorder. They're seeking insights into what encourages customer loyalty and repeated business.

To address this, we have employed a combination of feature engineering, data visualization, and decision trees in R to build a model that not only predicts the likelihood of customer churn but also helps in understanding the key determinants of customer retention.

Moreover, based on the insights derived from our model and supported by academic research on customer retention and churn prevention, we will recommend actionable strategies that the company could adopt to enhance customer loyalty and reduce the rate of customer churn